chor(styles): refactor style locations, imports and exports
