+++ src/components/About.tsx
@@ -1,10 +1,8 @@
-import React, { useState } from "react";
-import styled, { keyframes } from "styled-components";
+// Npm imports
+import React, { useState } from 'react';
+// Style imports
+import { AboutContainer, ContentInfoContainer } from '../styles';
 
-interface Props {
-    top: string,
-    order: number
-}
 
 const About = () => {
     let fileOrder = 1
@@ -34,68 +32,5 @@ const About = () => {
         </>
     )
 }
-
-const AboutContainer = styled.div`
-    z-index: 10;
-    background-image: url("folder.png");
-    background-size: cover;
-    background-position: center;
-    width: 9em;
-    height: auto;
-    padding-top: 2em;
-    
-    user-select: none;
-    -moz-user-select: none;
-    -webkit-user-select: none;
-    -ms-user-select: none;
-
-    & h2{
-        padding-left: 0.5em;
-    }
-`
-const enlarge = (x:number) => keyframes`
-    0%{
-        width: 0em;
-        height: .1em;
-        padding: 0;
-        left: 10em;
-    }
-    50%{
-        width: 10em;
-        height: .1em;
-        padding: 0;
-        left: (15 + (${x} * 4))em;
-    }
-    100%{
-        width: 30em;
-        height: 18em;
-        padding: 2em 1em;
-        left: (20 + (${x} * 4))em;
-    }
-`
-export const ContentInfoContainer = styled.div<Props>`
-    border: .1em solid black;
-    border-radius: 1em;
-    position: absolute;
-    top: ${props => props.top};
-    left: ${props => (props.order * 4) + 10}em; // Have the file move by two em to the right depending on the number of file clicked
-    z-index: ${props => props.order}; // Have the z-index increase by one depending on the numbers of files open
-    background-color: white;
-    overflow: hidden;
-    animation: ${props => enlarge(props.order)} 1.5s forwards ease-in-out;
-
-    & div{
-        position: absolute;
-        top: 1em;
-        right: 1.25em;
-        font-weight: bold;
-
-        // No text select
-        user-select: none;
-        -moz-user-select: none;
-        -webkit-user-select: none;
-        -ms-user-select: none;
-    }
-`
 
 export default About;
