+++ src/components/Taskbar.tsx
@@ -1,9 +1,17 @@
+// Npm imports
 import React, { useState, useEffect } from 'react';
-import styled from 'styled-components';
+// Image imports
 import logo from "../assets/logo.png";
 import bluetooth from "../assets/bluetooth.png";
 import wifi from "../assets/wifi.png";
 import speaker from "../assets/speaker.png";
+// Style imports
+import { 
+    TaskbarContainer,
+    LeftContainer,
+    RightContainer,
+    TimeContainer
+} from '../styles';
 
 const Taskbar = () => {
 
@@ -40,49 +48,5 @@ const Taskbar = () => {
     )
 }
 
-const TaskbarContainer = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    height: 2.75em;
-    background-color: #00000088;
-    position: fixed;
-    bottom: 0;
-    left: 0;
-    right: 0;
-`
-const LeftContainer = styled.div`
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    border-radius: 55px;
-    font-size: 2em;
-    color: white;
-
-    & img{
-        padding: 0 0.25em;
-        height: auto;
-        width: 1em;
-    }
-`
-const RightContainer = styled.div`
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-
-    & img{
-        height: auto;
-        width: 1.5em;
-        padding-right: 0.75em;
-    }
-`
-const TimeContainer = styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    color: white;
-    padding-right: 1.25em;
-`
 
 export default Taskbar;
