+++ src/styles/index.tsx
@@ -1,0 +1,22 @@
+import { AboutContainer } from './AboutStyles';
+import { ContactContainer } from './ContactStyles';
+import { ContentInfoContainer } from './ContentInfoStyles';
+import { ProjectsContainer } from './ProjectsStyles';
+import {
+    TaskbarContainer,
+    LeftContainer,
+    RightContainer,
+    TimeContainer
+} from './TaskbarStyles';
+
+export {
+    AboutContainer,
+    ContactContainer,
+    ContentInfoContainer,
+    ProjectsContainer,
+    TaskbarContainer,
+    LeftContainer,
+    RightContainer,
+    TimeContainer
+}
+
